+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- HUD position settings are now stored in `config/justcoordinates.toml` instead of JSON
+
 ## [0.9.0] - 2026-09-19
 
 ### Added